move authentications to separate file
