+++ enclave-runtime/src/tls_ra/authentication.rs
@@ -1,0 +1,118 @@
+/*
+	Copyright 2021 Integritee AG and Supercomputing Systems AG
+
+	Licensed under the Apache License, Version 2.0 (the "License");
+	you may not use this file except in compliance with the License.
+	You may obtain a copy of the License at
+
+		http://www.apache.org/licenses/LICENSE-2.0
+
+	Unless required by applicable law or agreed to in writing, software
+	distributed under the License is distributed on an "AS IS" BASIS,
+	WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+	See the License for the specific language governing permissions and
+	limitations under the License.
+
+*/
+
+//! Implementation of the client part of the state provisioning.
+
+use crate::cert;
+use itp_ocall_api::EnclaveAttestationOCallApi;
+use log::*;
+use sgx_types::*;
+use webpki::DNSName;
+
+pub struct ClientAuth<A> {
+	outdated_ok: bool,
+	skip_ra: bool,
+	attestation_ocall: A,
+}
+
+impl<A> ClientAuth<A> {
+	pub fn new(outdated_ok: bool, skip_ra: bool, attestation_ocall: A) -> Self {
+		ClientAuth { outdated_ok, skip_ra, attestation_ocall }
+	}
+}
+
+impl<A> rustls::ClientCertVerifier for ClientAuth<A>
+where
+	A: EnclaveAttestationOCallApi,
+{
+	fn client_auth_root_subjects(
+		&self,
+		_sni: Option<&DNSName>,
+	) -> Option<rustls::DistinguishedNames> {
+		Some(rustls::DistinguishedNames::new())
+	}
+
+	fn verify_client_cert(
+		&self,
+		_certs: &[rustls::Certificate],
+		_sni: Option<&DNSName>,
+	) -> Result<rustls::ClientCertVerified, rustls::TLSError> {
+		debug!("client cert: {:?}", _certs);
+		// This call will automatically verify cert is properly signed
+		if self.skip_ra {
+			warn!("Skip verifying ra-report");
+			return Ok(rustls::ClientCertVerified::assertion())
+		}
+
+		match cert::verify_mra_cert(&_certs[0].0, &self.attestation_ocall) {
+			Ok(()) => Ok(rustls::ClientCertVerified::assertion()),
+			Err(sgx_status_t::SGX_ERROR_UPDATE_NEEDED) =>
+				if self.outdated_ok {
+					warn!("outdated_ok is set, overriding outdated error");
+					Ok(rustls::ClientCertVerified::assertion())
+				} else {
+					Err(rustls::TLSError::WebPKIError(webpki::Error::ExtensionValueInvalid))
+				},
+			Err(_) => Err(rustls::TLSError::WebPKIError(webpki::Error::ExtensionValueInvalid)),
+		}
+	}
+}
+
+pub struct ServerAuth<A> {
+	outdated_ok: bool,
+	skip_ra: bool,
+	attestation_ocall: A,
+}
+
+impl<A> ServerAuth<A> {
+	pub fn new(outdated_ok: bool, skip_ra: bool, attestation_ocall: A) -> Self {
+		ServerAuth { outdated_ok, skip_ra, attestation_ocall }
+	}
+}
+
+impl<A> rustls::ServerCertVerifier for ServerAuth<A>
+where
+	A: EnclaveAttestationOCallApi,
+{
+	fn verify_server_cert(
+		&self,
+		_roots: &rustls::RootCertStore,
+		certs: &[rustls::Certificate],
+		_hostname: webpki::DNSNameRef,
+		_ocsp: &[u8],
+	) -> Result<rustls::ServerCertVerified, rustls::TLSError> {
+		debug!("server cert: {:?}", certs);
+
+		if self.skip_ra {
+			warn!("Skip verifying ra-report");
+			return Ok(rustls::ServerCertVerified::assertion())
+		}
+
+		// This call will automatically verify cert is properly signed
+		match cert::verify_mra_cert(&certs[0].0, &self.attestation_ocall) {
+			Ok(()) => Ok(rustls::ServerCertVerified::assertion()),
+			Err(sgx_status_t::SGX_ERROR_UPDATE_NEEDED) =>
+				if self.outdated_ok {
+					warn!("outdated_ok is set, overriding outdated error");
+					Ok(rustls::ServerCertVerified::assertion())
+				} else {
+					Err(rustls::TLSError::WebPKIError(webpki::Error::ExtensionValueInvalid))
+				},
+			Err(_) => Err(rustls::TLSError::WebPKIError(webpki::Error::ExtensionValueInvalid)),
+		}
+	}
+}
